+++ x/deployment/genesis.go
@@ -11,6 +11,7 @@ import (
 	"pkg.akt.dev/go/node/deployment/v1beta4"
 
 	"pkg.akt.dev/node/x/deployment/keeper"
+	"pkg.akt.dev/node/x/deployment/keeper/keys"
 )
 
 // ValidateGenesis does validation check of the Genesis and return error in case of failure
@@ -33,13 +34,20 @@ func DefaultGenesisState() *v1beta4.GenesisState {
 
 // InitGenesis initiate genesis state and return updated validator details
 func InitGenesis(ctx sdk.Context, kpr keeper.IKeeper, data *v1beta4.GenesisState) {
-	cdc := kpr.Codec()
-	store := ctx.KVStore(kpr.StoreKey())
+	k := kpr.(*keeper.Keeper)
 
 	for _, record := range data.Deployments {
-		key := keeper.MustDeploymentKey(keeper.DeploymentStateToPrefix(record.Deployment.State), record.Deployment.ID)
-
-		store.Set(key, cdc.MustMarshal(&record.Deployment))
+		pk := keys.DeploymentIDToKey(record.Deployment.ID)
+		has, err := k.Deployments().Has(ctx, pk)
+		if err != nil {
+			panic(fmt.Errorf("deployment genesis init. deployment id %s: %w", record.Deployment.ID, err))
+		}
+		if has {
+			panic(fmt.Errorf("deployment genesis init. deployment id %s: %w", record.Deployment.ID, v1.ErrDeploymentExists))
+		}
+		if err := k.Deployments().Set(ctx, pk, record.Deployment); err != nil {
+			panic(fmt.Errorf("deployment genesis init. deployment id %s: %w", record.Deployment.ID, err))
+		}
 
 		for idx := range record.Groups {
 			group := record.Groups[idx]
@@ -48,8 +56,10 @@ func InitGenesis(ctx sdk.Context, kpr keeper.IKeeper, data *v1beta4.GenesisState
 				panic(v1.ErrInvalidGroupID)
 			}
 
-			gkey := keeper.MustGroupKey(keeper.GroupStateToPrefix(group.State), group.ID)
-			store.Set(gkey, cdc.MustMarshal(&group))
+			gpk := keys.GroupIDToKey(group.ID)
+			if err := k.Groups().Set(ctx, gpk, group); err != nil {
+				panic(fmt.Errorf("deployment genesis groups init. group id %s: %w", group.ID, err))
+			}
 		}
 	}
 
@@ -60,23 +70,34 @@ func InitGenesis(ctx sdk.Context, kpr keeper.IKeeper, data *v1beta4.GenesisState
 }
 
 // ExportGenesis returns genesis state for the deployment module
-func ExportGenesis(ctx sdk.Context, k keeper.IKeeper) *v1beta4.GenesisState {
+func ExportGenesis(ctx sdk.Context, k keeper.IKeeper) (*v1beta4.GenesisState, error) {
 	var records []v1beta4.GenesisDeployment
-	k.WithDeployments(ctx, func(deployment v1.Deployment) bool {
-		groups := k.GetGroups(ctx, deployment.ID)
 
+	err := k.WithDeployments(ctx, func(deployment v1.Deployment) bool {
 		records = append(records, v1beta4.GenesisDeployment{
 			Deployment: deployment,
-			Groups:     groups,
 		})
 		return false
 	})
+	if err != nil {
+		return nil, err
+	}
+
+	for i := range records {
+		var groups v1beta4.Groups
+		groups, err = k.GetGroups(ctx, records[i].Deployment.ID)
+		if err != nil {
+			return nil, err
+		}
+
+		records[i].Groups = groups
+	}
 
 	params := k.GetParams(ctx)
 	return &v1beta4.GenesisState{
 		Deployments: records,
 		Params:      params,
-	}
+	}, nil
 }
 
 // GetGenesisStateFromAppState returns x/deployment GenesisState given raw application