+++ x/deployment/module.go
@@ -171,13 +171,16 @@ func (am AppModule) InitGenesis(ctx sdk.Context, cdc codec.JSONCodec, data json.
 // ExportGenesis returns the exported genesis state as raw bytes for the deployment
 // module.
 func (am AppModule) ExportGenesis(ctx sdk.Context, cdc codec.JSONCodec) json.RawMessage {
-	gs := ExportGenesis(ctx, am.keeper)
+	gs, err := ExportGenesis(ctx, am.keeper)
+	if err != nil {
+		panic(fmt.Errorf("deployment ExportGenesis failed: %w", err))
+	}
 	return cdc.MustMarshalJSON(gs)
 }
 
 // ConsensusVersion implements module.AppModule#ConsensusVersion
 func (am AppModule) ConsensusVersion() uint64 {
-	return 5
+	return 6
 }
 
 // AppModuleSimulation functions